Revert filterset change
The definition of filtersets in amsterdam schema was changed.
However, this was done for a test and the implementation in the filters.
So, existing datasets were crashing (at django startup).

The schema change was not very sensible, introducing a "kwargs" keyword
in amsterdam schema and using (and being dependent on) snake_case attributes.

So, this change has been reverted.