Make the CSV/GeoJSON response truely streaming.
The hacks to invoke a streaming response were nullified by the eventual
read of response.content. This instead, the whole DRF Response object is
replaced with an actual StreamingHttpResponse that Django handles
properly. The DRF logic is layered on top to maintain the
'accepted_render' logic to invoke the desired output renderer.