FIX using camelCase table names vs feature names in WFS
The use of underscore - as a result of camelCase table names -
was not taken into account when building the WFS viewset.

This led to a 400 Bad Request error. Which was a result of not finding the
table in the internal models due to the ignore of table names with underscores.