bug fix AUTH specification on table
In the permissions.py the method `fetch_scopes_for_dataset_table` used the table id as specified in the JSON schema as lookup value for DatasetTable object to get the objects auth attribute. Due to camelCase the object could not be found. The object was present in it's snake_case format. As a fix the lookup was added with a snake_case translation. In the future a general approach of using snake_case and camelCase in the dso-api logic could be revisited.