AB#5505 Make OpenCensus config dependant on `CLOUD_ENV`.
OpenCensus configuration is Azure specific. Hence will only be done if
the environment variable `CLOUD_ENV` is set to an azure specific
value (see README.md).

Don't use OpenCensus for logging purposes as each log message would then
result in an HTTP request. This has the potential to slow down the
application massively. Instead keep logging to `stdout` and rely on
external systems to redirect the log messages on `stdout` to the
appropriate log aggregator.

In addition require `AZURE_INGESTION_ENDPOINT` to be specified. Without
it a non-sensical default value will be used that pretty much guarantees
that we will not see any logging and trace information in Azure.