Pass 'fields_to_expand' logic to nested layers of serializers.
This builds on top of the refactoring done earlier which makes changes
the data structures in such way this information is available in the
first place.

The tricky part are M2M relations, where a nested DSOListSerializer
holds a next indirection level of DSOSerializer. The top-level list
performs it's own embedding, while in nested lists delegate that work to
the child serializer.