API docs: order of tables are presented by order of appearance in schema definition
On request of datateam omgevingswet the presentation order of the tables within a dataset is not ordered on alphabet but on appearance within the Amsterdam schema specification. The order of the tables defined in the Amsterdam schema is based on a certain logical and intentional meaning. In respect of the self defined order of tables the datateams can use the docs more intuatieve.