AB#4378 Show additional temporal fields in through table
For many to many relations that are temporal, we need to add extra
fields on the relation itself, and not on the target of the relation.

Before this PR, is was not possible to show these extra fields.
The relation was represented by a hyperlink related field that was
showing the url to the target object and some additional temporal fields
of the target object.

A new serializer, the ThroughSerializer, has been introduced that
represents the throught table, sitting between the source and the target
in a many to many relation.

This serializer needs to show:

- the urls + temporal compound key fields on the target (as before)
- the extra temporal field on the through table (usually begin- and
eindGeldigheid)

The amsterdam schema provides information about the temporal fields of a
table. This information is used to construct these fields for
the throughSerializer.

To show the information about the target (url + temporal compound key
fields), the already existing HALTemporalHyperlinkedRelatedField is
being used. However, the output of this field is not show as a subfield,
but merged into the fields of the ThroughSerializer.