AB#11744 browserview data loads with ajax
The html template for the api Browserview now loads the data
with a separate ajax call to the JSON version of its own endpoint.

This makes the frontend appear a lot  more responsive, as the head of the
page loads faster.

All the JSON printing logic has now moved clientside.