Fix OpenAPI layouts for temporal links and M2M model representations.
Instead of adjusting dicts in `to_representation()`, the serializer
layout can already reflect the output beforehand. This makes the code
more natural (no hacks to_representation(), just read properties) and it
causes OpenAPI spec to be generated correctly from this internal layout.

One effect is the removal of `obj.is_temporal()` checks at runtime. The
relations are either created as `DSORelatedLinkField`, or a special
`_temporal_link_serializer_factory()` field is generated. The same class
is both used for the `self` links and foreign objects of `_links`.

It also turned out to be possible to generate some related links based
on the foreignkey_ID field alone, reducing the number of queries.