Remove docker-compose.yml database from instructions
Our installation instructions mention that we have a database setup in
docker-compose.yml, then warn that it shouldn't be used, then proceed to
explain how to use it anyway. Just point to the dataservices-airflow one
instead and show a four-liner on how to run it.

Aside: f65c299b02de1919cd2461cfbce65140dea12079 removed the last mention
of dso_database, which also isn't mentioned in the dataservices-airflow
docs. This commit restores it.