AB#24678 Fix test_views_api.py
The auth tests assumed they would get 200 when filtering on a field that
doesn't exist. That conflicts with the requirement of getting a 400 in
this case. Add the field ("volgnummer") and check for 404 instead, since
we can't deliver the 400 just yet.

There's a test that requires the same behavior for
"regimes.inWerkingOp", which was removed for AB#24402. This still needs
updating.