Don't default to test JWKS (AB#38587) (#482)
* Don't default to test JWKS (AB#38587)

The test key belongs in src/tests/settings.py, not the main settings.py.
It also needs to be set when developing on the command line, or DSO-API
fails to start with an AuthzConfigurationError. This is inconvenient,
but much safer than the old setup.

* Separate token generation and request in auth docs

Since maketoken is a management command, running it loads all of Django
and takes several seconds. Errors from the command are also easier to
spot this way.

* Expand documentation on tokens and keysets