Restore mandatory filter bypass in permissions.py
Partial revert of 8609348380047381e14c14c31ad498814ca7eb54. I removed
too much of the mandatoryFilters handling, breaking the test.

The issue here is that when a filter is mandatory, it must also be
allowed, regardless of whether the field being used to filter is
accessible. This may be something that we need to rethink, but for now,
this test needs to pass.