Implemented reverse-relation support in the new filter engine
This replaces some functions to use a FilterPathPart that will handle
both a DatasetFieldSchema and AdditionalRelationSchema object. These
objects differ tremendously in their API and usage.

The early requests permissions check are removed, as the filters already do this.
This also avoids double-parsing the request, and double bookkeeping
when new features are added to the filter engine.