Request only the fields that the client accesses
This avoids requesting fields for which access is denied
(so database rules can be added). It also works when the ?_fields=..
parameter is given, as it checks which fields the serializer will read.

The prefetch-related queries are also updated to fetch a limited set of
fields. The resolve_model_lookup() now returns the field instead of the
model, since both field, related_model and is_many fields are needed.