Reorganize LinksSerializer so it longer has overrides to disable functionality.
Previously, the DynamicLinksSerializer had to disable the "_fields" and
"_expand" parameter logic that it inherited from the base class.
Now it only inherits the logic it actually needs: field access and field
construction.

The field-access logic is separated into a mixin that various classes can reuse.
This also allows reusing that logic for the elements inside the _links field later.

The DSOModelSerializer is also split in a part that does field
contruction, and the request-specific handling.