Implement field access checks for items in `_links`.
Previously the items in `_links` would be exposed, which may give access
to the display field, or identifiers for which access is denied.

This is achieved by letting the new FieldAccessMixin perform checks for
those serializers. Instead of letting those items extend from
ModelSerializer, the factory functions create an LinkSerializer object
instead.

As side effect, the 'self' link was also affected, which was fixed by
making `fields_always_included` more prominent.

The unit tests were affected, because the "afval" dataset did in fact
limit permissions on one relation.

Lastly, when a link item would not return any fields,
it will omit the entry completely.