[AB#50108] Loose relations: perform temporal filtering on loose relations
This makes sure the loose relations are receiving the proper temporal
filtering, as they'd return all records otherwise. This change is
possible now that the LooseRelationField is an actual ForeignKey, that
is properly setup to join on its main identifier field.

The TemporalTableQuery object no longer needed a request object to
perform the job. This is now used by QueryFilterEngine to apply this
logic within the filter parsing.