Simplify filter implementation
An empty django.db.models.Q object is a no-op and is False in a boolean
context. Use that to our advantage: we don't need to build up lists of
Q's with special handling for empty lists. Only
QueryFilterEngine.filter_queryset needs to check whether it needs to do
any work.