gzip all the responses
This reduces bandwidth for ordinary JSON responses by a factor 4-7,
GeoJSON by up to a factor 10, and even MVT responses are compressible.

HTTP compression exposes us to the BREACH attack, which reconstructs
secrets in the response body by looking at how its length changes from
echoing parts of the requests. In practice, that means CSRF tokens can
be leaked. But we don't use CSRF tokens for anything. We pass secrets
in the request header, which doesn't get compressed.