Add API Key Middleware
For now, the API Key has been set to be optional.
However, if provided, it will be validated.

pip-tools needed to be upgraded to get it running again.
Django check gave a warning about the Middleware not being in the
correct order in de settings.MIDDLEWARE, so it has been moved.