AB#101824 Temporary fix because schematools cannot be upgraded
DSO is nog pinned on Django 3.x, however, schematools has been
pinned on Dango >= 4.2. So, atm. schematools (with new functionality
that is needed) cannot be used.

For now, a fix has been made by adding a function with
the new functionality in schematools.

When schematools can be upgraded again, this commit can be removed.