AB#101824 The sealing check of deferred access bites us
Some access of the database is going directly to the DSO layer
of the DSO API and does not hit the dynamic_api layer, that
handles the amsterdam (schema) specific auth issues.

At the moment, the sealing raises an error. This has been made
configurable for now, so that access to deferred attributes
only issues a warning and return None (so not accessing the database,
because that could still lead to db permission error.)

More investigation is needed if there is a way to intercept
requests in the dynamic_api layer.