AB#107499 Fix problem with embedded serializer fetching all columns
Errors raised during streaming can never lead to a 403, because
when streaming, the response has already started with a status 200.

So, before the response starts, the validity of the request
should already have been checked and a 403 produced if needed.

Added tests to verify behavior.

Add tricks to the tests to make sure code is touched
by flushing prefetch caches.