Improve early database error handling, it happened too early
Moving beyond the last page should give a 404. Due to our generator
approach, this functionality had to be restored. However, instead of
letting that error arise during the actual pagination checks, it was
moved into __init__(). That give some extra expected errors on top of
errors.

By moving the logic into a second event, this code path is much cleaner,
and iteration does actually happen at the moment it should be.