Update schema tools to fix mock data (#930)
* Update schema tools to fix mock data

When running docker compose with mock data the command failed due to an
outdated schema being provided. We now use the active Amsterdam schema
to generate mock data.

The provided schema container is now also obsolute and thus removed.

* Use local schemas in openapi spec test

Starting up the container with the full Amsterdam Schema takes to long,
so we use the schemas container for this test using a specific docker
compose override.

The schemas have been updated to a recent version.