Fix bug where mainGeometry is not added as the default feature
When there are multiple geometries defined in the schema and the
mainGeometry comes after another geometry, it was not added as the
default feature. When the other geometry has required permission, this
caused an improperlyconfigured error.

This fix makes sure the maingeometry is added as the default/first
feature in wfs.