Re-implement chain to return generator instead of iterator
This was the cause of csv output not being streamed because
itertools.chain returns an iterator and not a generator. We test for the
type generator in our serializer to return a ReturnGenerator type
instead of a ReturnList type. This caused the whole csv to be generated
in memory before streaming the output. This fix makes sure output is
streamed while generating the csv.